requirements.txt
anthropic==0.7.4
openai==1.2.0
flask==2.3.3
transformers==4.34.0
torch==2.1.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0
pytest==7.4.2
python-dotenv==1.0.0
tqdm==4.66.1
flask-cors==4.0.0
requests==2.31.0
datasets==2.14.5